add setup for running the DSRM P1 telegram reader in a Docker container
To run you need Docker running on your Raspberry Pi (for example HypriotOS)
and run `docker-compose up` from the command line on the RPI. This will
trigger a build of the image and run a container after the build.

To quit hit Ctrl+c and the container will stop. To remove the container
run `docker-compose down`. To run again just run `docker-compose up` and
the container will start without doing the build again.